add excel dan config and bug siswacontroller

+++ database/migrations/2023_09_26_110216_create_config_table.php
@@ -33,7 +33,7 @@ return new class extends Migration
             $table->string('midtrans_merchant_id');
             $table->string('midtrans_client_key');
             $table->string('midtrans_server_key');
-            $table->bigInteger('order_id_midtrans')->nullable();
+            $table->integer('order_id_midtrans')->nullable();
             // $table->string('redirect_wa')->nullable();
             $table->timestamps();
         });
